파일삭제 완 서식자료실 자료 다운로드,  각 페이지 권한부여 해결못한 것 공지사항 알람 Y일 시 오류

+++ src/main/java/com/wisehr/wisehr/dataFormat/repository/DataFormatRepository.java
@@ -4,6 +4,6 @@ import com.wisehr.wisehr.dataFormat.entity.DataFormat;
 import com.wisehr.wisehr.notice.entity.Notice;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DataFormatRepository extends JpaRepository<DataFormat, Integer> {
+public interface DataFormatRepository extends JpaRepository<DataFormat, Long> {
 
 }
